Initial import
This is an initial import of some basic code I've been playing around
with.

The importer and example_search scripts both work. I have the output
in the search showing the score, I'm hoping this will provide some
interesting information but not sure it does just yet ( was hoping the
score would go closer to 1 based on distance )